Rename mod to state_machine
